+++ src/Problems/BBOB/f_gallagher21.hpp
@@ -62,7 +62,7 @@ public:
   
   void prepare_problem() {
     const long rseed = (long) (22 + 10000 * this->IOHprofiler_get_instance_id());
-    int n = IOHprofiler_get_number_of_variables();
+    size_t n = (size_t) IOHprofiler_get_number_of_variables();
 
     double fopt;
     size_t i, j, k;
@@ -160,7 +160,7 @@ public:
 
 
   double internal_evaluate(const std::vector<double> &x) {
-    int n = x.size();
+    size_t n = x.size();
 
     size_t i, j; /* Loop over dim */
     std::vector<double> tmx(n);
