+++ src/Problems/BBOB/f_katsuura.hpp
@@ -69,7 +69,7 @@ public:
   }
 
   double internal_evaluate(const std::vector<double> &x) {
-    int n = x.size();
+    size_t n = x.size();
     size_t i, j;
     double tmp, tmp2;
     std::vector<double> result(1);
